+++ src/main/java/net/congstar/jira/plugins/scrumpoker/action/ShowSessionsAction.java
@@ -2,11 +2,15 @@ package net.congstar.jira.plugins.scrumpoker.action;
 
 import com.atlassian.jira.issue.IssueManager;
 import com.atlassian.jira.issue.MutableIssue;
+import com.atlassian.jira.permission.ProjectPermissions;
+import com.atlassian.jira.security.JiraAuthenticationContext;
+import com.atlassian.jira.security.PermissionManager;
 import com.atlassian.jira.user.util.UserManager;
 import net.congstar.jira.plugins.scrumpoker.data.ScrumPokerStorage;
 import net.congstar.jira.plugins.scrumpoker.model.ScrumPokerSession;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class ShowSessionsAction extends ScrumPokerAction {
 
@@ -15,11 +19,16 @@ public class ShowSessionsAction extends ScrumPokerAction {
     private final ScrumPokerStorage scrumPokerStorage;
     private final UserManager userManager;
     private final IssueManager issueManager;
+    private final PermissionManager permissionManager;
+    private final JiraAuthenticationContext jiraAuthenticationContext;
 
-    public ShowSessionsAction(ScrumPokerStorage scrumPokerStorage, UserManager userManager, IssueManager issueManager) {
+    public ShowSessionsAction(ScrumPokerStorage scrumPokerStorage, UserManager userManager, IssueManager issueManager,
+                              PermissionManager permissionManager, JiraAuthenticationContext jiraAuthenticationContext) {
         this.scrumPokerStorage = scrumPokerStorage;
         this.userManager = userManager;
         this.issueManager = issueManager;
+        this.permissionManager = permissionManager;
+        this.jiraAuthenticationContext = jiraAuthenticationContext;
     }
 
     @Override
@@ -28,11 +37,17 @@ public class ShowSessionsAction extends ScrumPokerAction {
     }
 
     public List<ScrumPokerSession> getOpenSessions() {
-        return scrumPokerStorage.getOpenSessions();
+        return scrumPokerStorage.getOpenSessions().stream()
+                .filter(session -> permissionManager.hasPermission(ProjectPermissions.BROWSE_PROJECTS,
+                        issueManager.getIssueObject(session.getIssueKey()), jiraAuthenticationContext.getLoggedInUser()))
+                .collect(Collectors.toList());
     }
 
     public List<ScrumPokerSession> getClosedSessions() {
-        return scrumPokerStorage.getClosedSessions();
+        return scrumPokerStorage.getClosedSessions().stream()
+                .filter(session -> permissionManager.hasPermission(ProjectPermissions.BROWSE_PROJECTS,
+                        issueManager.getIssueObject(session.getIssueKey()), jiraAuthenticationContext.getLoggedInUser()))
+                .collect(Collectors.toList());
     }
 
     public String getUsername(String key) {